refactor(updates): extract shared release-age logic into @argent/update-core
The installer and tool-server each carried near-identical npm release-age
logic: config-value parsing, minimum-release-age detection, packument fetch,
and installable-target selection. Consolidate it into a new CommonJS
@argent/update-core package consumed by both.

- update-target.ts (installer) drops from ~240 lines to a thin wrapper.
- update-checker.ts (tool-server) drops its inlined fetch/pick/parse helpers;
  min-release-age.ts is deleted.
- detectMinReleaseAgeMsForPm (installer knows the PM) and detectMinReleaseAgeMs
  (server probes all, takes the max) are preserved as two exports sharing one
  probe table — the only genuine divergence between the two copies.
- Pure-logic unit tests now live once in update-core; each consumer keeps a
  thin wiring test mocking the @argent/update-core boundary.
