+++ packages/tool-server/package.json
@@ -16,6 +16,7 @@
     "@argent/native-devtools-ios": "file:../native-devtools-ios",
     "@argent/native-devtools-android": "file:../native-devtools-android",
     "@argent/registry": "file:../registry",
+    "@argent/update-core": "file:../update-core",
     "@clack/prompts": "^1.1.0",
     "express": "^4.19.2",
     "pngjs": "^7.0.0",
