+++ packages/tool-server/src/utils/update-checker.ts
@@ -1,14 +1,17 @@
 import fs from "node:fs";
-import https from "node:https";
 import os from "node:os";
 import path from "node:path";
 import semver from "semver";
 import { version as currentVersion } from "../../package.json";
-import { detectMinReleaseAgeMs } from "./min-release-age";
+import {
+  detectMinReleaseAgeMs,
+  fetchRegistryInfo,
+  pickInstallableTarget,
+} from "@argent/update-core";
 
 const PACKAGE_NAME = "@swmansion/argent";
+const REGISTRY_URL = `https://registry.npmjs.org/${PACKAGE_NAME}`;
 const CHECK_INTERVAL_MS = 60 * 60 * 1000 * 24; // 24 hour
-const REQUEST_TIMEOUT_MS = 10_000;
 
 /** Set truthy to silence the update reminder entirely (no check, no note). */
 const DISABLE_ENV = "ARGENT_DISABLE_UPDATE_NOTIFICATIONS";
@@ -109,116 +112,9 @@ function isNewerVersion(latest: string, current: string): boolean {
   return semver.gt(latest, current);
 }
 
-// No policy → everything passes. Under a policy, an unknown/unparseable
-// publish time conservatively returns false (delay rather than nag).
-function isOldEnough(publishedAt: string | null, minReleaseAgeMs: number): boolean {
-  if (minReleaseAgeMs <= 0) return true;
-  if (!publishedAt) return false;
-  const published = Date.parse(publishedAt);
-  if (Number.isNaN(published)) return false;
-  return Date.now() - published >= minReleaseAgeMs;
-}
-
-interface VersionAt {
-  version: string;
-  publishedAt: string | null;
-}
-
-/**
- * The newest stable version newer than `current` that the resolver could
- * install now — i.e. that also clears the policy. With no policy this is just
- * the latest tag; under one we scan all versions, since the latest publish may
- * be held while an older version is eligible. `null` when nothing is installable.
- */
-function pickInstallableTarget(
-  latest: VersionAt,
-  times: Record<string, string>,
-  current: string,
-  minReleaseAgeMs: number
-): VersionAt | null {
-  if (minReleaseAgeMs <= 0) {
-    return isNewerVersion(latest.version, current) ? latest : null;
-  }
-
-  let best: VersionAt | null = null;
-  for (const [version, publishedAt] of Object.entries(times)) {
-    // `times` also carries non-version keys ("created"/"modified") — filtered as invalid semver.
-    if (!semver.valid(version) || semver.prerelease(version)) continue;
-    if (!semver.gt(version, current)) continue;
-    if (!isOldEnough(publishedAt, minReleaseAgeMs)) continue;
-    if (best === null || semver.gt(version, best.version)) {
-      best = { version, publishedAt };
-    }
-  }
-  return best;
-}
-
-interface RegistryInfo {
-  latest: VersionAt;
-  /** version → ISO 8601 publish time (also includes created/modified). */
-  times: Record<string, string>;
-}
-
-// Fetch the full packument (not /latest) — only it carries the `time` map the
-// release-age gate needs. Returns null on any failure (never crashes the server).
-async function fetchRegistryInfo(): Promise<RegistryInfo | null> {
-  return new Promise((resolve) => {
-    let resolved = false;
-    const safeResolve = (value: RegistryInfo | null) => {
-      if (!resolved) {
-        resolved = true;
-        resolve(value);
-      }
-    };
-
-    const url = `https://registry.npmjs.org/${PACKAGE_NAME}`;
-
-    const req = https.get(url, { timeout: REQUEST_TIMEOUT_MS }, (res) => {
-      if (res.statusCode !== 200) {
-        res.resume();
-        safeResolve(null);
-        return;
-      }
-
-      let body = "";
-      res.setEncoding("utf8");
-      res.on("data", (chunk: string) => {
-        body += chunk;
-      });
-      res.on("end", () => {
-        try {
-          const json = JSON.parse(body) as {
-            "dist-tags"?: { latest?: string };
-            "time"?: Record<string, string>;
-          };
-          const latestVersion = json["dist-tags"]?.latest;
-          if (!latestVersion) {
-            safeResolve(null);
-            return;
-          }
-          const times = json.time ?? {};
-          safeResolve({
-            latest: { version: latestVersion, publishedAt: times[latestVersion] ?? null },
-            times,
-          });
-        } catch {
-          safeResolve(null);
-        }
-      });
-      res.on("error", () => safeResolve(null));
-    });
-
-    req.on("error", () => safeResolve(null));
-    req.on("timeout", () => {
-      req.destroy();
-      safeResolve(null);
-    });
-  });
-}
-
 /** Run a single check and update the runtime state. */
 async function check(): Promise<void> {
-  const info = await fetchRegistryInfo();
+  const info = await fetchRegistryInfo(REGISTRY_URL);
   if (info === null) return; // network issue — keep previous state
 
   const minReleaseAgeMs = await detectMinReleaseAgeMs();
