+++ packages/update-core/package.json
@@ -1,25 +1,18 @@
 {
   "private": true,
-  "name": "@argent/installer",
+  "name": "@argent/update-core",
   "version": "0.8.1",
-  "description": "Workspace setup logic for argent: init, update, uninstall",
-  "type": "module",
+  "description": "Shared npm release-age detection and installable-target resolution for argent updates",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "rm -rf dist tsconfig.tsbuildinfo && tsc",
     "test": "vitest run",
+    "test:watch": "vitest",
     "typecheck:tests": "tsc --noEmit -p tsconfig.test.json"
   },
   "dependencies": {
-    "@argent/tools-client": "file:../argent-tools-client",
-    "@argent/update-core": "file:../update-core",
-    "@clack/prompts": "^1.4.0",
-    "jsonc-parser": "^3.3.1",
-    "picocolors": "^1.1.1",
-    "semver": "^7.8.0",
-    "smol-toml": "^1.6.1",
-    "yaml": "^2.9.0"
+    "semver": "^7.7.4"
   },
   "devDependencies": {
     "@types/node": "^25.9.0",
