+++ packages/update-core/tests/min-release-age.test.ts
@@ -4,11 +4,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 const { mockExec } = vi.hoisted(() => ({ mockExec: vi.fn() }));
 vi.mock("node:child_process", () => ({ exec: mockExec }));
 
-import {
-  detectMinReleaseAgeMs,
-  parseBeforeConfigValue,
-  parseConfigValue,
-} from "../src/utils/min-release-age";
+import { detectMinReleaseAgeMs, detectMinReleaseAgeMsForPm } from "../src/min-release-age";
 
 const NOW = new Date("2026-06-01T00:00:00Z");
 const DAY_MS = 24 * 60 * 60 * 1000;
@@ -30,64 +26,19 @@ function stubPmConfig(byCommand: Record<string, string>) {
   });
 }
 
-describe("parseConfigValue", () => {
-  it("treats unset markers as no policy (0)", () => {
-    expect(parseConfigValue("undefined")).toBe(0);
-    expect(parseConfigValue("null")).toBe(0);
-    expect(parseConfigValue("")).toBe(0);
-    expect(parseConfigValue("   \n")).toBe(0);
-  });
-
-  it("parses positive numbers and ignores surrounding whitespace", () => {
-    expect(parseConfigValue("1440\n")).toBe(1440);
-    expect(parseConfigValue("  7 ")).toBe(7);
-    expect(parseConfigValue('"30"')).toBe(30);
-  });
-
-  it("rejects non-positive and non-finite values", () => {
-    expect(parseConfigValue("0")).toBe(0);
-    expect(parseConfigValue("-5")).toBe(0);
-    expect(parseConfigValue("Infinity")).toBe(0);
-    expect(parseConfigValue("not-a-number")).toBe(0);
-  });
+beforeEach(() => {
+  vi.useFakeTimers();
+  vi.setSystemTime(NOW);
+  mockExec.mockReset();
+  delete process.env[ENV];
 });
 
-describe("parseBeforeConfigValue", () => {
-  it("treats unset markers as no policy (0)", () => {
-    expect(parseBeforeConfigValue("undefined", NOW.getTime())).toBe(0);
-    expect(parseBeforeConfigValue("null", NOW.getTime())).toBe(0);
-    expect(parseBeforeConfigValue("", NOW.getTime())).toBe(0);
-  });
-
-  it("converts npm's effective `before` date into an equivalent age", () => {
-    const twoDaysAgo = new Date(NOW.getTime() - 2 * DAY_MS).toISOString();
-    expect(parseBeforeConfigValue(twoDaysAgo, NOW.getTime())).toBe(2 * DAY_MS);
-  });
-
-  it("parses npm's human-readable date output with a trailing timezone label", () => {
-    const value = "Fri Apr 03 2026 00:00:12 GMT+0200 (Central European Summer Time)";
-    expect(parseBeforeConfigValue(value, Date.parse("2026-04-05T00:00:12+02:00"))).toBe(2 * DAY_MS);
-  });
-
-  it("returns 0 for invalid or future cutoff dates", () => {
-    expect(parseBeforeConfigValue("not-a-date", NOW.getTime())).toBe(0);
-    expect(parseBeforeConfigValue("2030-01-01T00:00:00Z", NOW.getTime())).toBe(0);
-  });
+afterEach(() => {
+  vi.useRealTimers();
+  delete process.env[ENV];
 });
 
-describe("detectMinReleaseAgeMs", () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
-    vi.setSystemTime(NOW);
-    mockExec.mockReset();
-    delete process.env[ENV];
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-    delete process.env[ENV];
-  });
-
+describe("detectMinReleaseAgeMs (probe every PM)", () => {
   it("returns 0 when no package manager reports a policy", async () => {
     stubPmConfig({
       "npm config get before": "null",
@@ -145,6 +96,27 @@ describe("detectMinReleaseAgeMs", () => {
     process.env[ENV] = "not-a-number";
     expect(await detectMinReleaseAgeMs()).toBe(0);
 
+    expect(mockExec).not.toHaveBeenCalled();
+  });
+});
+
+describe("detectMinReleaseAgeMsForPm (known PM)", () => {
+  it("probes only the selected package manager", async () => {
+    const sevenDaysAgo = new Date(NOW.getTime() - 7 * DAY_MS).toISOString();
+    stubPmConfig({ "npm config get before": sevenDaysAgo });
+    expect(await detectMinReleaseAgeMsForPm("npm")).toBe(7 * DAY_MS);
+    expect(mockExec).toHaveBeenCalledTimes(1);
+  });
+
+  it("returns 0 for bun (no `config get`, covered by the override instead)", async () => {
+    stubPmConfig({});
+    expect(await detectMinReleaseAgeMsForPm("bun")).toBe(0);
+    expect(mockExec).not.toHaveBeenCalled();
+  });
+
+  it("honors the override and skips probing", async () => {
+    process.env[ENV] = "2";
+    expect(await detectMinReleaseAgeMsForPm("pnpm")).toBe(2 * DAY_MS);
     expect(mockExec).not.toHaveBeenCalled();
   });
 });
