+++ packages/update-core/tsconfig.json
@@ -6,12 +6,6 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "references": [
-    { "path": "../registry" },
-    { "path": "../update-core" },
-    { "path": "../native-devtools-ios" },
-    { "path": "../native-devtools-android" }
-  ],
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
 }
